FIX: halt overcrowded message now displayed again
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@2202 8aca7d54-2c30-db11-9de9-000461428c89

+++ gui/message_stats_t.cc
@@ -35,7 +35,7 @@ void message_stats_t::infowin_event(const event_t * ev)
 {
 	if(IS_LEFTRELEASE(ev)) {
 		const int line = (ev->cy - 15) / 14;
-		if(line<msg->gib_count()) {
+		if((unsigned)line<msg->gib_count()) {
 			message_t::node *n=msg->get_node(line);
 			if(ev->cy>14  &&  n!=NULL  &&  welt->ist_in_kartengrenzen(n->pos)) {
 				welt->change_world_position(koord3d(n->pos,welt->min_hgt(n->pos)));
